requirements.txt
Django>=3.1
psycopg2
python-dateutil